Adds an edit command
