+++ Package.resolved
@@ -1,0 +1,70 @@
+{
+  "object": {
+    "pins": [
+      {
+        "package": "Files",
+        "repositoryURL": "https://github.com/JohnSundell/Files.git",
+        "state": {
+          "branch": null,
+          "revision": "6171f9e9d8619678da22355560f3dbed6368058d",
+          "version": "2.0.1"
+        }
+      },
+      {
+        "package": "Marathon",
+        "repositoryURL": "https://github.com/JohnSundell/Marathon.git",
+        "state": {
+          "branch": "edit_edit",
+          "revision": "8fdf8014d1ce7fc50ea4b0933d324d6d1366b50d",
+          "version": null
+        }
+      },
+      {
+        "package": "Releases",
+        "repositoryURL": "https://github.com/JohnSundell/Releases.git",
+        "state": {
+          "branch": null,
+          "revision": "0a3c1ba8ce8bb27c02c225979bfd02888b38f0cc",
+          "version": "2.0.1"
+        }
+      },
+      {
+        "package": "Require",
+        "repositoryURL": "https://github.com/JohnSundell/Require.git",
+        "state": {
+          "branch": null,
+          "revision": "7cfbd0d8a2dede0e01f6f0d8ab2c7acef1df112e",
+          "version": "2.0.1"
+        }
+      },
+      {
+        "package": "ShellOut",
+        "repositoryURL": "https://github.com/JohnSundell/ShellOut.git",
+        "state": {
+          "branch": null,
+          "revision": "479a89d9bfa7585496bd4a3e4faffac5b01af37f",
+          "version": "2.0.0"
+        }
+      },
+      {
+        "package": "Unbox",
+        "repositoryURL": "https://github.com/JohnSundell/Unbox.git",
+        "state": {
+          "branch": null,
+          "revision": "17ad6aa1cc2f1efa27a0bbbdb66f79796708aa95",
+          "version": "2.5.0"
+        }
+      },
+      {
+        "package": "Wrap",
+        "repositoryURL": "https://github.com/JohnSundell/Wrap.git",
+        "state": {
+          "branch": null,
+          "revision": "8085c925060b84a1fc0caef444c130da2c8154c3",
+          "version": "3.0.1"
+        }
+      }
+    ]
+  },
+  "version": 1
+}
